fix(indentation): Resolve incorrect indentation after a function call
This commit fixes the behaviour where no semicolon is inserted after a
closing parentheses e.g. in a function call func() and the cursor goes
to the next line with an extra indentation.

The changes were made to the `setter.lua` file to check for a closing
parentheses. This reduces the need of pressing on the space bar, then
enter, in order to insert the semicolon.